// src/types/media.ts
export type MediaType =
  | "featured_image"
  | "normal_images"
  | "demo_video"
  | "youtube_video"
  | "logo"
  | "profile_pic"
  | "community_discussion_image"
  | "community_discussion_video"
  | "community_discussion_comment_image"
  | "community_discussion_comment_video"
  | "forum_image"
  | "forum_video"
  | "forum_comment_image"
  | "forum_comment_video"
  | "post_image"
  | "post_video"
  | "post_comment_video"
  | "post_comment_image"
  | "submission_image"
  | "submission_video"
  | "submission_comment_image"
  | "submission_comment_video"
  | "announcement_image"
  | "announcement_video"
  | "changelog_image"
  ;
export type MediaContentType = "video" | "image";